Apply styles to wrapper elements for form components
+ change it to a reusable composable
+ update all imports to use the @ alias instead of path
